replace static_casts with c-style casts
because they are all about casting integers and floats

+++ src/geometry/bezier.cpp
@@ -8,16 +8,16 @@ BezierCurve::BezierCurve(int bezier_segments) : bezier_segments(bezier_segments)
 }
 
 void BezierCurve::DrawControlPoints(Color color_points, Color color_lines) const {
-    for (int i = 1; i < static_cast<int>(control_points.size()); ++i) {
+    for (int i = 1; (size_t) i < control_points.size(); ++i) {
         DrawLineDotted(control_points[i - 1], control_points[i], 20, 3, color_lines);
     }
-    for (int i = 0; i < static_cast<int>(control_points.size()); ++i) {
+    for (int i = 0; (size_t) i < control_points.size(); ++i) {
         DrawCircleV(control_points[i], 7, color_points);
     }
 }
 
 void BezierCurve::DrawCurve(Color color) const {
-    for (int i = 1; i < static_cast<int>(curve_points.size()); ++i) {
+    for (int i = 1; (size_t) i < curve_points.size(); ++i) {
         DrawLineV(curve_points[i - 1], curve_points[i], color);
     }
 }
@@ -32,7 +32,7 @@ void BezierCurve::Update() {
 
     curve_points.clear();
     for (int i = 0; i <= bezier_segments; ++i) {
-        float t = static_cast<float>(i) / bezier_segments;
+        float t = (float) i / bezier_segments;
         curve_points.push_back(bezier_func(t));
     }
 }
